+++ sdk/metricsadvisor/ai-metrics-advisor/src/models.ts
@@ -209,7 +209,7 @@ export interface DataFeedOptions {
   rollupSettings?: DataFeedRollupSettings;
 
   /**
-   * settings to control how missing data apoints are filled
+   * settings to control how missing data points are filled
    */
   missingDataPointFillSettings?: DataFeedMissingDataPointFillSettings;
 
@@ -270,13 +270,13 @@ export interface DataFeed {
   /**
    * Time when the data feed is created
    */
-  createdTime: Date;
+  createdOn: Date;
   /**
    * Status of the data feed.
    */
   status: DataFeedStatus;
   /**
-   * Indicates whether the current user is an aministrator of the data feed.
+   * Indicates whether the current user is an administrator of the data feed.
    */
   isAdmin: boolean;
   /**
@@ -402,6 +402,14 @@ export type MongoDBDataFeedSource = {
 };
 
 /**
+ * Represents an Unknown data source.
+ */
+export type UnknownDataFeedSource = {
+  dataSourceType: "Unknown";
+  dataSourceParameter: unknown;
+};
+
+/**
  * Represents a SQL Server data source.
  */
 export type SQLServerDataFeedSource = {
@@ -425,7 +433,8 @@ export type DataFeedSource =
   | MySqlDataFeedSource
   | PostgreSqlDataFeedSource
   | SQLServerDataFeedSource
-  | MongoDBDataFeedSource;
+  | MongoDBDataFeedSource
+  | UnknownDataFeedSource;
 
 /**
  * Represents the input type to the Update Data Feed operation.
@@ -530,7 +539,7 @@ export type MetricSeriesGroupDetectionCondition = DetectionConditionsCommon & {
 };
 
 /**
- * Detection condidtion for a specific time series.
+ * Detection condition for a specific time series.
  */
 export type MetricSingleSeriesDetectionCondition = DetectionConditionsCommon & {
   /**
@@ -574,7 +583,7 @@ export type HardThresholdConditionUnion =
        */
       anomalyDetectorDirection: "Both";
       /**
-       * supress condition
+       * suppress condition
        */
       suppressCondition: SuppressCondition;
     };
@@ -635,7 +644,7 @@ export interface MetricFeedbackCommon {
   /**
    * feedback created time
    */
-  readonly createdTime?: Date;
+  readonly createdOn?: Date;
   /**
    * user who gives this feedback
    */