+++ sdk/metricsadvisor/ai-metrics-advisor/CHANGELOG.md
@@ -45,7 +45,9 @@
   - `EnrichmentStatus.timestamp`
   - `IngestionStatus.timestamp`
   - `latestSuccessTimestamp` and `latestActiveTimestamp` in the return type of `getDataFeedIngestionProgress()`.
+- [Breaking] property `createdTime` on `DataFeed` and `MetricFeedbackCommon` to `createdOn`.
 - Parameters of `Date` type now also accept strings. No validation is done for the strings. The SDK calls `new Date()` to convert them to `Date`.
+- Handle potential new data feed source types gracefully
 
 ## 1.0.0-beta.1 (2020-10-07)